move routes types with navigators
